Chat Implementation
Attempt to add chat features.
