+++ app/src/main/java/sg/edu/np/mad_p03_group_gg/tools/FirebaseTools.java
@@ -255,73 +255,5 @@ public  class FirebaseTools {
      * @param uID
      * @param message
      */
-    public static void sendConfirmationMessage(String sID, String uID, String message) {
-
-        databaseReference.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                String chatKey = null;
-
-                // Get chatkey
-                for (DataSnapshot dataSnapshotCurrentChat : snapshot.child("chat").getChildren()){
-                    // Get id number of each user
-                    String getUserOne = dataSnapshotCurrentChat.child("user1").getValue(String.class);
-                    String getUserTwo = dataSnapshotCurrentChat.child("user2").getValue(String.class);
-
-                    // If id numbers are the same as main user and selected user's id number
-                    if((TextUtils.equals(getUserOne,sID) && TextUtils.equals(getUserTwo,uID))
-                            || (TextUtils.equals(getUserOne,uID) && TextUtils.equals(getUserTwo, sID))){
-                        chatKey = dataSnapshotCurrentChat.getKey();
-                    }
-                }
-
-                // Setting chat key
-                if(chatKey == null) {
-                    // ChatKey increment by 1 for each chat. Default chat key is 1 (for first 2 users)
-                    chatKey = "1";
-
-                    if (snapshot.hasChild("chat")) {
-                        chatKey = String.valueOf(snapshot.child("chat").getChildrenCount() + 1);
-                    }
-                }
-
-                Boolean inChat = Boolean.parseBoolean(snapshot.child("chat").child(chatKey).child(sID).child("inChat").getValue(String.class));
-
-                // Get current time (add 28800000 milliseconds to convert to SGT, if emulator timezone is UTC)
-                String currentTime = String.valueOf(System.currentTimeMillis());
-
-                DatabaseReference ref = databaseReference.push();
-
-                // If user inchat status is true, set message seen value to True
-                if (inChat){
-                    ref.child("chat").child(chatKey).child("messages").child(currentTime).child("seen").setValue("True");
-                }
-                // If other user is not in current chat, set value to false
-                else{
-                    ref.child("chat").child(chatKey).child("messages").child(currentTime).child("seen").setValue("False");
-                }
-
-                // Set users
-                ref.child("chat").child(chatKey).child("user1").setValue(uID);
-                ref.child("chat").child(chatKey).child("user2").setValue(sID);
-
-                // Set message and who sent the message
-                ref.child("chat").child(chatKey).child("messages").child(currentTime).child("msg").setValue(message);
-                ref.child("chat").child(chatKey).child("messages").child(currentTime).child("id").setValue(uID);
-
-
-                // If current user (you) are not already in other user's friend list, add to his friend list
-                ref.child("selectedChatUsers").child(sID).child(uID).setValue("");
-
-                // Set in chat status to true if message is sent
-                ref.child("chat").child(chatKey).child(uID).child("inChat").setValue("True");
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-                // Failed to read from db
-            }
-        });
-    }
 
 }
